requirements.txt
gunicorn==20.0.4
Jinja2==2.11.2
Werkzeug==1.0.0
MarkupSafe==1.1.1
requests==2.25.0
Flask==1.1.1
flask-mysqldb